Add initial TPM implementation
We generate a new PrimaryKey for each registration from a random seed
and the application parameter. From the PrimaryKey we generate a child
P256 key. The resulting key handle is the encrypted key handle of this
child key, plus the random seed used to generate the PrimaryKey.

On auth requests we re-derive the PrimaryKey from the application
param + seed and then attempt to load the child key from the key
handle. If the application param, seed or key handle is different we
won't be able to sign the request.