Improve pinentry experience
- Fix some confirmations being lost
- Fixed timeout handling

If the browser stopped polling we were not dismissing the prompt after
2 seconds because we were trying to send to a channel that would never
be read from. We now timeout after 2 seconds if nothing reads from
that response channel so we properly dismiss the pinentry prompt.

We were also having confirmations being lost because of timing issues.
We were unsetting the activeRequest obj before writing to the response
channel. Depending on the state of the browser, this could result in
the accept click never getting delivered back to the application
logic. We now send to the result chan before clearing the
activeRequest object to avoid that race with the user.