Fix pinentry deadlock
Sometimes we'd get stuck where we'd stop responding to tpm messages as
we'd expect. This was easiest to reproduce by canceling a request and
then triggering a new request.

The deadlock was caused by ConfirmPresense locking the mutex that
protects activeRequest and then trying to send to the extendTimeout
channel while holding that lock. Its possible for nothing to be
reading from extendTimeout if we've gotten a result to send back.

When prompt is shutting down it will no longer read from the
extendTimeout chan and it needs to lock that same mutex to unset
activeRequest.

We now send to the extendTimeout chan in a goroutine so as to not
continue to hold the lock. We also will timeout that send incase the
reader has gone away.