Reduce the size our keyhandles generated
KeyHandles can't be larger than 255 bytes. On machines I had tested
with they came in under 255 bytes, but we've now identified other
systems where they can be larger.

Trim out the following fat:
- Switch to a custom lencode implementation that uses a 1 byte len
prefix instead of 4.
- Cut the size of the record separator from 7 bytes to 3. Having the
record separator was very helpful in debugging this issue so I don't
want to remove it completely.
- Cut the rand bytes from 32 to 20 bytes. The rand bytes is
important for us to not generate the same key if we register multiple
times (say for different user names). It should not affect the
security properties of the implementation (the TPM is protecting the
key).